+++ CHANGELOG.md
@@ -5,7 +5,15 @@
 ### Added
 
 * Add interactive mode for text command
-* Add parameter `history`, `model`, `temp`, `system` for text command
+* Add parameter for text command
+    * `sdk` : Select the sdk to use
+    * `history` : Select the history to use
+    * `model` : Select the model to use
+    * `temp` : Set the temperature
+    * `system` : Prompt with system role
+    * `file` : Add content of file text to history
+    * `clear` : Clear the history
+    * `list-history` : List all history
 * Add history with name for text command
 * Add sdk claude for text command
 * Add sdk openai for text command